Copy in the Sakila sample database; write code to use it
