+++ pgdo-lib/Cargo.toml
@@ -35,6 +35,7 @@ version = "^0.7.2"
 features = ["macros", "postgres", "runtime-tokio"]
 
 [dev-dependencies]
+async-std = "^1.12.0"
 paste = "^1.0.14"
 pgdo-test = { path = "../pgdo-test" }
 rand = "^0.8.5"
