+++ sakila/README.pgdo.md
@@ -1,0 +1,5 @@
+# The Sakila example database in pgdo
+
+The code in this folder is copied from [sakila](https://github.com/jOOQ/sakila)
+commit 38c2e4ef53cf7ef2eec873d42c5a1e2c7afa0231. The copy contains only the
+PostgreSQL parts, the upstream README.md, and the upstream LICENSE.
